Include README.md & RELEASE.md files in docs, make README first page
As the ex_doc CLI does not support the :extras option, I had to move the
CLI args to a config file. It would probably be simpler to add ex_doc as
a non-runtime dependency to this project, and then pass the arguments to
the docs directly from mix.exs, but I kept the existing escript setup.